requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
langchain-community
bs4
pypdf
chromadb
faiss-cpu
wikipedia
groq
cassio
langchain-groq
agent
numexpr
langchain_experimental
lamini
kagglehub
transformers
hf_xet
accelerate>=0.26.0